refactor($compile): 상품 Bounded Context 중심의 패키지 구조로 변경
