+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -1,7 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.products.application;
 
 import kitchenpos.domain.*;
 import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
