+++ src/main/java/kitchenpos/products/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.products.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
