+++ src/main/java/kitchenpos/products/domain/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.products.domain;
 
 import java.util.List;
 import java.util.Optional;
