+++ src/test/java/kitchenpos/products/application/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.products.application;
 
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
